Working client+server streaming support
Significanly simplified the request/response API so that everything for
the request from BatchScheduler->Client->ServerInstance is a serialized
plan, and the response back is an AsyncGenerator. Client and
BatchScheduler are now more decoupled. Unfortunately this broke the
existing batch scheduler tests and fixing them doesn't seem worth the
effort.

Realized there is an issue with the client API now that it's possible to
await on outputs and streams at the same time. Consider this:

```
const x = client.api.foo()
const stream = client.api.stream()
for await (const v of stream) { .. } // <-- batch finished here!
console.log(await x) // <-- this will fail!
```

The problem is that the batch scheduler finalizes the current batch on
the first await. So the second await will be called on the next batch.

Also problematic is lazy get operations. Thinking of adding a .defer()
method or something similar that eagerly calls pushOutput() on proxies
the client wants to resolve the outputs of after streams have been
drained.